Allow calibration of run before ratio calculated
Another ratio can be used to calibrate the denominator of the ratio
calculation, on a per OV level. The calibrated array needs to be the
same size as the two plots and cover the same range of OVs.

In the future this could be generalized to support numerator or
denominator, and a constant factor.

It should also probably be moved to the initialization of the
AnalyzeResults class, not during the plotting.